requirements.txt
Flask==2.3.3
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
requests==2.31.0
tensorflow==2.16.1
tf_keras==2.16.0
tensorflow_decision_forests==1.9.0